Move data classes to data folder
